Add & Edit orders fixed
añadir-pedido.php & editar-pedido.php fixed, now it shows proper message explaining if there's an error or if it's success.

+++ backend/api/pedidos/editar.php
@@ -3,46 +3,57 @@ session_start();
 require_once "../db.php";
 // echo '<pre>'; print_r($_POST); echo '</pre>';
 
-// 1. Datos del Cliente
-$pedido_id = $_GET['id'];
-$cliente_id = trim(mb_strtoupper($_POST['nombre'], 'UTF-8'));
-$forma_pago = trim(mb_strtoupper($_POST['pago'], 'UTF-8'));
-$fecha_estimada = trim(mb_strtoupper($_POST['fecha'], 'UTF-8'));
-$estado = 'EN ANALISIS';
+// Si se ejecuta un Request, ya sea GET o POST se ejecuta el código.
+if ($_SERVER["REQUEST_METHOD"] == "POST") {
 
-$sql = "UPDATE PEDIDOS SET CLIENTE_ID = ?, FECHA_ESTIMADA = ?, FORMA_PAGO = ? WHERE ID = ?;";
-$data = array($cliente_id, $fecha_estimada, $forma_pago, $pedido_id);
+    $pedido_id = $_GET['id'];
+    $pedidos = $_POST['pedido'];
+    $estado = 'EN ANALISIS';
 
-db_query($sql, $data);
-
-$sql = "DELETE FROM PRODUCCION WHERE PEDIDO_ID = ?;";
-db_query($sql, array($pedido_id));
-
-// 2. Datos del Pedido
-$pedidos = $_POST['pedido'];
-
-foreach ($pedidos as $pedido) {
+    $cliente_id = test_input($_POST['nombre']);
+    $forma_pago = test_input($_POST['pago']);
+    $fecha_estimada = test_input($_POST['fecha']);
     
-    if ($pedido['cantidad'] != 0){
+    $sql = "SELECT * FROM PEDIDOS WHERE ID = ?;";
+    $estado_actual = db_query($sql, array($pedido_id))[0]['ESTADO'];
 
-        if (isset($pedido['urgente'])) {
-            $urgente = trim(mb_strtoupper($pedido['urgente'], 'UTF-8'));
-        } else {
-            $urgente = 0;
+    if($estado_actual == 'EN ANALISIS'){
+
+        $sql = "UPDATE PEDIDOS SET CLIENTE_ID = ?, FECHA_ESTIMADA = ?, FORMA_PAGO = ? WHERE ID = ?;";
+        db_query($sql, array($cliente_id, $fecha_estimada, $forma_pago, $pedido_id));
+    
+        $sql = "DELETE FROM PRODUCCION WHERE PEDIDO_ID = ?;";
+        db_query($sql, array($pedido_id));
+    
+        foreach ($pedidos as $pedido) {
+            
+            if ($pedido['cantidad'] != 0){
+    
+                if (isset($pedido['urgente'])) {
+                    $urgente = test_input($pedido['urgente']);
+                } else {
+                    $urgente = 0;
+                }
+    
+                $suela_id = test_input($pedido['suela_id']);
+                $serie_id = test_input($pedido['serie_id']);
+                $color_id = test_input($pedido['color_id']);
+                $cantidad = test_input($pedido['cantidad']);
+                $restante = test_input($pedido['cantidad']);
+    
+                $sql = "INSERT INTO PRODUCCION VALUES (NULL, ?, ?, ?, ?, ?, ?, 0, 0, 0, 0, 0, ?, ?, NOW())";
+                db_query($sql, array($pedido_id, $suela_id, $serie_id, $color_id, $cantidad, $restante, $urgente, $estado));
+    
+            }
+    
         }
 
-        $suela_id = trim(mb_strtoupper($pedido['suela_id'], 'UTF-8'));
-        $serie_id = trim(mb_strtoupper($pedido['serie_id'], 'UTF-8'));
-        $color_id = trim(mb_strtoupper($pedido['color_id'], 'UTF-8'));
-        $cantidad = trim(mb_strtoupper($pedido['cantidad'], 'UTF-8'));
-        $restante = trim(mb_strtoupper($pedido['cantidad'], 'UTF-8'));
+        echo 'SUCCESS';
 
-        $sql = "INSERT INTO PRODUCCION VALUES (NULL, ?, ?, ?, ?, ?, ?, 0, 0, 0, 0, 0, ?, ?, NOW())";
-        $data = array($pedido_id, $suela_id, $serie_id, $color_id, $cantidad, $restante, $urgente, $estado);
-        db_query($sql, $data);
+    } else {
+
+        echo 'ERROR';
 
     }
-
+    
 }
-
-header("Location: ../../../pedidos-pendientes.php");
